requirements.txt
gspread==6.1.0
oauth2client==4.1.3
pandas==2.2.1
google-auth==2.29.0
psycopg2-binary
fastapi
uvicorn
orjson